libnarrator : rolling narrator
narrator which writes to a set of files described by a file size threshold for
rollover and maximum number of files. Files share a common path and differ by a
numeric suffix. Upon rollover, if the file already exists, it is first unlinked.
Unlinking rather than truncating an existing file makes it easier for an
application to tail these logs, say with inotify.

Signed-off-by: Todd Freed <todd.freed@gmail.com>