fabd: simplify path and move from liblorien to fabd
This is the first step in unraveling all of the path canonicalization, which is
too expensive. Instead, path now only does normalization, which is much simpler
and requires no io.

This will have two important consequences. First, the --gnid feature whereby
paths are transformed on display to the user is no longer possible. It has not
that useful, and clutters the ux besides. Second, paths to the same object must
now be given with the same normalized name, that is, symlinks will not be
resolved. This is a great performance boost, but it is possible that this
decision will need to be revisited. On possibility would be a customizable level
of path canonicalization specified in the filesystem config. But it will be best
to avoid this complexity entirely.

Signed-off-by: Todd Freed <todd.freed@gmail.com>