listwise: fab will no longer use listwise, so remove it from the repository
In the v1 fabfile syntax, listwise is used for three things.

1. Constructing variables (like $CC, $CFLAGS, etc).
2. Reflecting over the dependency graph for interpolation into formula text.
3. Inspecting the filesystem to identify primary files, and their derivatives.

It is awkaward and inefficient for all of these tasks. They will be implemented
differently in v2 fabfile syntax.

1. Such an expressive syntax is overkill here. What is necessary is to be able
to define families of related variables, e.g. variant variables. But variables
should be defined declaratively.

2. Overkill again. Its only ever necessary to do this in a couple of different
ways. Passing variables to the formula with the list of dependencies will
suffice here.

3. IO has historically been the performance bottleneck for fab, and filesystem
access is being streamlined in the rewrite. The in memory dependency graph will
always reflect the state of the filesystem, and logic for constructing secondary
files will not touch the filesystem.

listwise has been extracted to another repository
https://github.com/hossbeast/listwise

Signed-off-by: Todd Freed <todd.freed@gmail.com>