liblogger/expr: move expr parsing out of filter and into expr
exprs can now include attrs as well as filters

liblogger/attrs: rename PROCESSID -> NAMES

thread name is already supported, in addition to process name

liblogger/category: optional categories

When an optional category is specified in a filter, the filter can only match
logs which include that category. A filter need not include all of the optional
categories on a log to match that log.

liblogger/log: logger_trace, log a backtrace with xapi_trace

liblogger/log: simplify start/finish to use an intermediary narrator rather than
a token

When starting a log message, if not log_would, then log_start returns the
nullity narrator.

The log in progress is appended to using the narrator apis, rather than the log
apis with category_ids of zero

Signed-off-by: Todd Freed <todd.freed@gmail.com>