liblogger: defer merging category lists until activate
This patch fixes a bug in the previous algorithm where successively registering
and activating lists of categories would incorrectly detect an ordering
incompatibility, because the intermediary lists contained implicit dependencies
which did not exist in the original lists.

Instead of merging lists two at a time, the new algorithm retains the source
lists, and merges them in a single operation.

Signed-off-by: Todd Freed <todd.freed@gmail.com>