build: dont use modules from the previous fab release
fab includes builtin modules for building certain common file types.

However, because fab is built by the previous release of fab, and
because the fab source contains files in the class of common file types
for which builtin modules are provided, fab itself cannot use the
builtin modules. To do so would lead to a circular dependency, and it
would be necessary to make an intermediate release any time those files
are changed.

To avoid this situation, make the build logic self contained, by
including two sets of modules, 1) the builtin modules that fab provides,
and 2) the modules that fab itself builds with.

Note that modules in #1, are to be used by the previous release of fab,
while modules in #2 are used by the current version.

In addition, support for building fab with the buildscripts is removed
for now.

Signed-off-by: Todd Freed <todd.freed@gmail.com>