moria: graph: additional operations on an embedded identity tree
An identity tree is a prefix tree, and has the property that each vertex
is uniquely identified by its path to some vertex, called the root.

For nodes in the prefix tree, graph_lookup is a lookup operation with
logarithmic time complexity in the number of vertices.

The dependency graph used by fab contains an embedded identity tree and
all nodes that represent files are in it (RELATION_TYPE_FS).

In order to test these operations, this patch also introduces the
canonical graph representation (graph_say), and two parsers for
constructing graphs from a description.

The first parser (libmoria/parser) allows for constructing a graph from
its canonical representation.

The second parser (libmoria/operations) allows for assembling a sequence
of graph operations from a description. The operations can then be
executed on a subject graph to perform a series of mutations, including
direct construction, when the subject graph is empty.

Signed-off-by: Todd Freed <todd.freed@gmail.com>