value: merge: no error cases
Instead of having incompatible types and throwing errors during merge,
types are promoted during merge, including synthesis of an aggregate
when two scalars are merged.

value: clone: the clone operation

clone is a deep-copy operation which is used by the merge operation.

value: query: remove the query operation

query is no longer needed. Each place that parses bacon text (config
files ; requests ; formulas) has a dedicated parser that builds data
structures suitable for its purpose, so there is no longer any
intermediate format.

value: parse_partial

parse_partial allows for optimistic parsing of up to N bytes ; it's not
an error if the entire buffer is not consumed. This is required of a
parser which is to be called in a nested fashion from another parser.

value: writer

For writing formatted bacon messages

value: value (executable)

Standalone executable for validating / formatting bacon messages

Signed-off-by: Todd Freed <todd.freed@gmail.com>