fabd: the rewrite
This patch is the culmination of a complete rewrite of fabd, the core
component of this project.

High level summary of differences, compared to fab-v0.5:

Filesystem i/o is now completely centralized. The primary datastructure
in fabd is the graph, and embedded in the graph is an identity tree
representing the filesystem. The identity tree is constructed at
startup, and then maintained by processing inotify events, and also by
walking filesystem trees not watched by inotify, before processing each
request. Thus fabd is always operating on a consistent view of the
filesystem from some (recent) point in time. There is no filesystem
access except as necessary to keep the identity tree up to date - all
operations are against the graph.

fabfiles have been replaced with module.bam files which have a
completely different syntax. Notably, listwise expressions (which were
evaluated against the filesystem) are replaced with patterns/rules which
are evaluated against the graph.

There is a system for user configuration with system/user/project config
files and precedence.

Formulas as text in a fabfile containing variables which are
interpolated just before execution have been removed (this was a lot of
io). Formulas are now standalone executable files (scripts) which take
parameters as args and environment variables. They may contain a block
of bacon text at the end of the file describing parameters necessary for
execution, which fabd prepares.

Support for reusing state across consecutive executions as different
users is removed (the analog of make && sudo make install). This is not
going to be supported going forward. fabd simply runs as the user who
executed fab.

Support for dependency discovery is removed (for now).

Support for buildscripts is removed (for now).

Signed-off-by: Todd Freed <todd.freed@gmail.com>