fabd: build: dont trigger shutdown on exec failure
Fixes a bug in which a formula subprocess triggers program shutdown
after exec fails. This happens because following an exec failure
(because e.g. the file does not exist) the process proceeds to the build
thread teardown logic (since it was spawned from the build thread). Fix
by adding the missing exit after exec.

Signed-off-by: Todd Freed <todd.freed@gmail.com>