fabd: build: config for $PATH, support path-search
New configuration (under formula.path) for specification of $PATH to
pass to formula subprocesses. $PATH in the users env at the time fab is
run is only relevant when formula.path.copy-from-env is true.

The top-level formula/bacon key path is renamed to file. In addition to
taking a scalar, it can take the newly added path-search operation which
performs path lookup against $PATH as specified in formula.path
(equivalent to the similarly named operation performed by a shell). The
path-search operation may also be supplied as the value of a mapping in
the top-level envs set.

To support the usage of path-search, 1) the path_cache module is
introduced, and 2) formulas which specify the top-level file parameter
are now executed by file descriptor (with fexec) rather than by absolute
path.

Signed-off-by: Todd Freed <todd.freed@gmail.com>