narrator: refactor to use a base type
Changes struct narrator from an opaque type to a base type, which
implementing types embed as the first member. This technique makes makes
different types of different sizes simpler to use.

Removes the default N narrator. This was a piece of hidden context that
would be surprising to newer users.

Signed-off-by: Todd Freed <todd.freed@gmail.com>