libmoria: refactor to use embedded structs rather than opaque types
Previously, vertices and edges were allocated at a fixed, specified
size, and able to carry a value at the end of the allocation. Now,
vertices and edges are (from the perspective of libmoria) just as large
as they need to be. The application embeds them in another structure of
sufficient size to carry the value, if any.

This change makes it a lot more natural to have different objects of
different sizes embedded in a graph.