build: use build slot chaining to support execs with large strings
This patch addresses an issue that arises when the combined size of the
args, envs, sources, and targets strings for an exec exceeds the shm
page size.

fabd will now submit multiple pages for an exec, fab chains the pages
together and processes them as a list. Any given exec string must fit
within a single page.

Signed-off-by: Todd Freed <todd.freed@gmail.com>