libfab: ipc: actually awaken waiters
Fixes a problem with libfab/ipc in which the consume and post operations
did not correctly awaken waiters, which meant that waiters would often
stall until the 125ms timeout.

libfab: ipc: support many concurrent producers

Fixes an issue in which concurrent producers might produce at the index
of a previously produced and not yet consumed message, leading to
corrupted messages. A mechanism is added for producers beyond a fixed
max (16) to wait until other active producers have posted.

Signed-off-by: Todd Freed <todd.freed@gmail.com>