formula: values: define env var in the case of zero-entry list value
When a formula value resolves to a list with zero entries, the current
behavior is not to pass the environment variable to the formula at all.
This patch changes the behavior in this case to define the environment
variable with a zero-length value. This simplifies things for formula
authors because they can rely on the environment variable always at
least existing.

Signed-off-by: Todd Freed <todd.freed@gmail.com>